Polishing sistem

+++ tcc/src/main/java/com/br/uvv/tcc/usecase/ProductCrudUseCase.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Service;
 import com.br.uvv.tcc.entities.Product;
 import com.br.uvv.tcc.entities.enums.Role;
 import com.br.uvv.tcc.gateway.database.ProductDatabase;
+import com.br.uvv.tcc.usecase.exceptions.NotEnoughProductInStockBusinessException;
+import com.br.uvv.tcc.usecase.exceptions.ProductNotFoundBusinessException;
 
 @Service
 public class ProductCrudUseCase {
@@ -24,6 +26,10 @@ public class ProductCrudUseCase {
 	}
 	
 	public void sell(final Long id, final Integer quantitySold) {
+		final Product productSold = this.productDatabase.get(id);
+		if(productSold.getQuantity() - quantitySold < 0) {
+			throw new NotEnoughProductInStockBusinessException();
+		}
 		this.productDatabase.sell(id, quantitySold);
 	}
 	
@@ -33,6 +39,10 @@ public class ProductCrudUseCase {
 	
 	public void delete(final Long id, final Role role) {
 		this.verifyCredentialsUseCase.verifyManager(role);
+		final Product product = this.productDatabase.get(id);
+		if(product == null) {
+			throw new ProductNotFoundBusinessException();
+		}
 		this.productDatabase.delete(id);
 	}
